+++ app/uk/gov/hmrc/gform/binders/ValueClassBinder.scala
@@ -18,7 +18,7 @@ package uk.gov.hmrc.gform.binders
 
 import play.api.libs.json.{ JsError, JsString, JsSuccess, Reads }
 import play.api.mvc.PathBindable
-import uk.gov.hmrc.gform.models.form._
+import uk.gov.hmrc.gform.gformbackend.model.{ FormId, FormTypeId, Version }
 
 object ValueClassBinder {
