+++ app/uk/gov/hmrc/gform/controllers/SummaryGen.scala
@@ -21,9 +21,9 @@ import javax.inject.{ Inject, Singleton }
 import play.api.i18n.{ I18nSupport, MessagesApi }
 import play.api.libs.json.Json
 import uk.gov.hmrc.gform.controllers.helpers.FormDataHelpers._
+import uk.gov.hmrc.gform.gformbackend.model.{ FormId, FormTypeId, Version }
 import uk.gov.hmrc.gform.models._
 import uk.gov.hmrc.gform.models.components.FieldId
-import uk.gov.hmrc.gform.models.form._
 import uk.gov.hmrc.gform.service.SaveService
 import uk.gov.hmrc.play.frontend.controller.FrontendController
