+++ app/uk/gov/hmrc/gform/views/field_template_address.scala.html
@@ -1,7 +1,6 @@
 @import uk.gov.hmrc.gform.models.components.FieldValue
 @import uk.gov.hmrc.gform.models.FormFieldValidationResult
 @import uk.gov.hmrc.play.views.html.helpers._
-@import uk.gov.hmrc.gform.models.ComponentField
 
 @(international: Boolean, fieldValue: FieldValue, validationResult: Option[FormFieldValidationResult])
