+++ app/uk/gov/hmrc/gform/views/field_template_date.scala.html
@@ -1,6 +1,5 @@
+@import uk.gov.hmrc.gform.models.{FormFieldValidationResult, ValidationUtil}
 @import uk.gov.hmrc.gform.models.components.{DateExpr, FieldValue}
-@import uk.gov.hmrc.gform.models.FormFieldValidationResult
-@import uk.gov.hmrc.gform.models.ValidationUtil
 @import uk.gov.hmrc.play.views.html.helpers._
 
 @(fieldValue: FieldValue, validationResult: Option[FormFieldValidationResult], prepop: Option[DateExpr])
