+++ app/uk/gov/hmrc/gform/views/field_template_info.scala.html
@@ -1,7 +1,4 @@
-@import uk.gov.hmrc.gform.models.components.FieldValue
-@import uk.gov.hmrc.gform.models.components.{InfoType,StandardInfo,LongInfo,ImportantInfo,BannerInfo}
-@import uk.gov.hmrc.gform.models.FormFieldValidationResult
-@import uk.gov.hmrc.play.views.html.helpers._
+@import uk.gov.hmrc.gform.models.components.{BannerInfo, FieldValue, ImportantInfo, InfoType, LongInfo, StandardInfo}
 
 @(fieldValue: FieldValue, infoType: InfoType, text: Html)
