+++ app/uk/gov/hmrc/gform/views/field_template_text.scala.html
@@ -1,5 +1,5 @@
-@import uk.gov.hmrc.gform.models.components.{FieldValue, Text}
 @import uk.gov.hmrc.gform.models.FormFieldValidationResult
+@import uk.gov.hmrc.gform.models.components.{FieldValue, Text}
 @import uk.gov.hmrc.play.views.html.helpers._
 
 @(fieldValue: FieldValue, text: Text, prepop: String, validationResult: Option[FormFieldValidationResult])
