+++ app/uk/gov/hmrc/gform/views/file_upload.scala.html
@@ -1,6 +1,4 @@
 @import uk.gov.hmrc.gform.models.components.FieldValue
-@import uk.gov.hmrc.gform.models.FormFieldValidationResult
-@import uk.gov.hmrc.play.views.html.helpers._
 
 @(fieldValue: FieldValue)
