+++ app/uk/gov/hmrc/gform/views/group.scala.html
@@ -1,7 +1,4 @@
-@import uk.gov.hmrc.gform.models.components.{FieldValue, Orientation, Horizontal, Vertical}
-@import uk.gov.hmrc.play.views.html.helpers._
-@import views.html.helper.CSRF
-@import scala.concurrent.Future
+@import uk.gov.hmrc.gform.models.components.{FieldValue, Horizontal, Orientation}
 
 @(fieldValue: FieldValue, snippets: List[Html], orientation: Orientation)
