+++ app/uk/gov/hmrc/gform/views/hidden_field.scala.html
@@ -1,4 +1,4 @@
-@import uk.gov.hmrc.gform.models.form.FormField
+@import uk.gov.hmrc.gform.gformbackend.model._
 
 @(formField: FormField)
