+++ app/uk/gov/hmrc/gform/views/snippets/summary/address.scala.html
@@ -1,8 +1,6 @@
-@import uk.gov.hmrc.gform.models.components.FieldValue
 @import uk.gov.hmrc.gform.models.FormFieldValidationResult
+@import uk.gov.hmrc.gform.models.components.FieldValue
 @import uk.gov.hmrc.gform.views.html.localisation
-@import uk.gov.hmrc.gform.views._
-@import uk.gov.hmrc.play.views.html.helpers._
 
 @(fieldValue: FieldValue, validationResult: Option[FormFieldValidationResult])
