+++ app/uk/gov/hmrc/gform/views/snippets/summary/begin_section.scala.html
@@ -1,7 +1,5 @@
-@import uk.gov.hmrc.gform.models.form.{FormTypeId, FormId, Version}
+@import uk.gov.hmrc.gform.gformbackend.model.{FormId, FormTypeId, Version}
 @import uk.gov.hmrc.gform.views.html.localisation
-@import uk.gov.hmrc.play.views.html.helpers._
-
 @(formType: FormTypeId, version: Version, formId: FormId, title: String, idx: Int)
 
 <table class="table--font-reset @if(idx == 0) { table--top-border }">
