+++ app/uk/gov/hmrc/gform/views/snippets/summary/choice.scala.html
@@ -1,6 +1,5 @@
 @import uk.gov.hmrc.gform.models.components.FieldValue
 @import uk.gov.hmrc.gform.views.html.localisation
-@import uk.gov.hmrc.play.views.html.helpers._
 
 @(fieldValue: FieldValue, selections: List[String])
