+++ app/uk/gov/hmrc/gform/views/snippets/summary/date.scala.html
@@ -1,8 +1,7 @@
-@import uk.gov.hmrc.gform.models.components.FieldValue
 @import uk.gov.hmrc.gform.models.FormFieldValidationResult
-@import uk.gov.hmrc.gform.views.html.localisation
+@import uk.gov.hmrc.gform.models.components.FieldValue
 @import uk.gov.hmrc.gform.models.helpers.DateHelperFunctions._
-@import uk.gov.hmrc.play.views.html.helpers._
+@import uk.gov.hmrc.gform.views.html.localisation
 
 @(fieldValue: FieldValue, validationResult: Option[FormFieldValidationResult])
