+++ app/uk/gov/hmrc/gform/views/snippets/summary/end_section.scala.html
@@ -1,6 +1,4 @@
-@import uk.gov.hmrc.gform.models.form.{FormTypeId, FormId, Version}
-@import uk.gov.hmrc.gform.views.html.localisation
-@import uk.gov.hmrc.play.views.html.helpers._
+@import uk.gov.hmrc.gform.gformbackend.model.{FormId, FormTypeId, Version}
 
 @(formType: FormTypeId, version: Version, formId: FormId, title: String, currentPage: Int)
