+++ app/uk/gov/hmrc/gform/views/snippets/summary/text.scala.html
@@ -1,7 +1,6 @@
-@import uk.gov.hmrc.gform.models.components.{FieldValue,Text}
 @import uk.gov.hmrc.gform.models.FormFieldValidationResult
+@import uk.gov.hmrc.gform.models.components.{FieldValue, Text}
 @import uk.gov.hmrc.gform.views.html.localisation
-@import uk.gov.hmrc.play.views.html.helpers._
 
 @(fieldValue: FieldValue, text: Text, validationResult: Option[FormFieldValidationResult])
