+++ test/uk/gov/hmrc/gform/models/BooleanExprSpec.scala
@@ -17,8 +17,9 @@
 package uk.gov.hmrc.gform.models
 
 import org.scalatest._
+import uk.gov.hmrc.gform.gformbackend.model.{ FormTemplate, FormTypeId, Version }
 import uk.gov.hmrc.gform.models.components._
-import uk.gov.hmrc.gform.models.form._
+import uk.gov.hmrc.gform.gformbackend.model._
 
 class BooleanExprSpec extends FlatSpec with Matchers with EitherValues {
