+++ test/uk/gov/hmrc/gform/models/PageSpec.scala
@@ -18,8 +18,9 @@ package uk.gov.hmrc.gform.models
 
 import org.scalatest._
 import org.scalatest.concurrent.ScalaFutures
+import uk.gov.hmrc.gform.gformbackend.model.{ FormTemplate, FormTypeId, Version }
 import uk.gov.hmrc.gform.models.components.{ Constant, FieldId, FieldValue, Text }
-import uk.gov.hmrc.gform.models.form.{ FormTypeId, Version }
+import uk.gov.hmrc.gform.gformbackend.model._
 import uk.gov.hmrc.gform.models.helpers.Extractors.extractNames
 import uk.gov.hmrc.play.frontend.auth.AuthContext
 import uk.gov.hmrc.play.frontend.auth.connectors.domain.{ Accounts, Authority, ConfidenceLevel, CredentialStrength }
